+++ src/ble/alerts/NoHub.tsx
@@ -5,8 +5,8 @@ import './index.scss';
 import { AnchorButton, Button, Intent } from '@blueprintjs/core';
 import React from 'react';
 import { appName, pybricksBluetoothTroubleshootingUrl } from '../../app/constants';
+import ExternalLinkIcon from '../../components/ExternalLinkIcon';
 import type { CreateToast } from '../../toasterTypes';
-import ExternalLinkIcon from '../../utils/ExternalLinkIcon';
 import { useI18n } from './i18n';
 
 type NoHubProps = {
