+++ src/components/clipboard.scss
