+++ src/components/external-link.scss
