+++ src/sponsor/SponsorDialog.tsx
@@ -7,9 +7,9 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 import { alertsShowAlert } from '../alerts/actions';
 import { pybricksTeamUrl } from '../app/constants';
+import ClipboardIcon from '../components/ClipboardIcon';
+import ExternalLinkIcon from '../components/ExternalLinkIcon';
 import { useSelector } from '../reducers';
-import ClipboardIcon from '../utils/ClipboardIcon';
-import ExternalLinkIcon from '../utils/ExternalLinkIcon';
 import patreonLogo from './Digital-Patreon-Logo_White.png';
 import gitHubIcon from './GitHub-Mark-Light-32px.png';
 import { sponsorHideDialog } from './actions';
